separate message groups
