+++ src/shared/messages/tab.ts
@@ -1,0 +1,12 @@
+
+export type TabInfo = { tabId: number; windowId: number; };
+
+export type TabActivatedEvent = {
+  message?: never;
+  response: void;
+};
+
+export type GetTabInfoEvent = {
+  message?: never;
+  response: TabInfo;
+};
